# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: LightAI Go API
  version: "0.1.14"
  description: |
    LightAI Go is a lightweight GPU/node management platform.
    Base URL: http://localhost:18080/api/v1
servers:
  - url: http://localhost:18080/api/v1
    description: Local dev server
paths:
...
  /auth/change-password:
    post:
      operationId: changePassword
      summary: Change current user password
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password: { type: string }
                new_password: { type: string }
      responses:
        "200": { description: Password changed }

  # --- Nodes ---
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: lightai_session
    agentToken:
      type: http
      scheme: bearer